text_vector_service/requirements.txt
grpcio-tools~=1.67.1
grpcio~=1.67.1
langchain~=0.3.7
loguru~=0.7.2
pydantic~=2.9.2
pydantic-settings~=2.5.2
sentence-transformers~=3.2.1
qdrant-client[async]~=1.12.1
grpcio-reflection==1.67.1
mwparserfromhell~=0.6.6